+++ data-plane/core/tracing/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0](https://github.com/agntcy/slim/compare/slim-tracing-v0.2.7...slim-tracing-v0.3.0) - 2026-01-29
+
+### Added
+
+- *(bindings)* configuration file support ([#1099](https://github.com/agntcy/slim/pull/1099))
+
+### Other
+
+- unified typed error handling across core crates ([#976](https://github.com/agntcy/slim/pull/976))
+
 ## [0.2.7](https://github.com/agntcy/slim/compare/slim-tracing-v0.2.6...slim-tracing-v0.2.7) - 2025-11-21
 
 ### Other